Added validate.go to handle field validation
